requirements.txt
streamlit
pandas
langchain-openai
langchain-core
python-dotenv
feedparser
beautifulsoup4
plotly
scikit-learn
numpy